requirements.txt
fastapi
uvicorn
sqlalchemy
httpx
jinja2
python-multipart
python-dotenv
